Introduced limit of max. pending tasks when copying files
Previously, SharpRemote was flooded with as many RPCs as possible, leading to reading the entire file into memory faster than it was sent over the network. Now the number of pending tasks has been reduced to 20 which means that no more than 80Mb of a file are kept in memory at a time.